loading 1000 school entries to initialize dev environment with data.
