+++ config.yaml
@@ -6,7 +6,7 @@ server:
 
 database:
   driverClass: org.postgresql.Driver
-  url: jdbc:postgresql://localhost:5432/postgres
+  url: jdbc:postgresql://localhost:5432/school
   user: postgres
   password:
