+++ run-application.sh
@@ -1,4 +1,22 @@
 #!/usr/bin/env bash
-psql postgres -U postgres -c "CREATE TABLE IF NOT EXISTS ORDERED_ENTRY_INDEX (ID SERIAL PRIMARY KEY, ENTRY JSONB)"
-psql postgres -U postgres -c "CREATE TABLE IF NOT EXISTS STREAMED_ENTRIES (ID INTEGER PRIMARY KEY, TIME TIMESTAMP)"
+
+db_name=school
+
+createdb "$db_name"
+
+psql $db_name -U postgres -c "CREATE TABLE IF NOT EXISTS current_keys(key VARCHAR PRIMARY KEY, serial_number INTEGER UNIQUE)"
+psql $db_name -U postgres -c "CREATE TABLE IF NOT EXISTS total_records(count INTEGER)"
+
+psql $db_name -U postgres -c "insert into total_records (count) select 0 where not exists (select count from total_records)"
+
+psql $db_name -U postgres -c "CREATE TABLE IF NOT EXISTS ordered_entry_index(serial_number INTEGER PRIMARY KEY, entry JSONB)"
+psql $db_name -U postgres -c "CREATE TABLE IF NOT EXISTS total_entries(count INTEGER, last_updated TIMESTAMP WITHOUT TIME ZONE DEFAULT now())"
+
+psql $db_name -U postgres -c "insert into total_entries (count) select 0 where not exists (select count from total_entries)"
+
+
+if [ `psql $db_name -U postgres -c "select count(*) from ordered_entry_index" | egrep -v [a-z-]+ | grep -E '[0-9]' | sed 's/ //g'` -eq 0 ]; then
+    ./gradlew loadSchoolData
+fi
+
 ./gradlew run
